another restart
this time I was hoping to improve the coding style a little better.
with some ideas of how to structure things gotten from the first couple
iterations, I think this version is a bit cleaner. there is certainly
still room for improvement in this iteration as well, but hopefully I
cna clean this one up without another full restart.

so far, the reader works to read in symbols and lists*. it should not be
difficult to add strings to that, as well as expanding on the
classification of tokens so that numbers can be read in as well.

\* the list reading is limited at the moment; it cannot read in dotted
lists.