update allocation and deallocation logic
basic free / allocated lists, and pulling off first fit. one thing that
might make this serendipitously good is that a lot of objects will
probably be temporary, so recently allocated objects are put at the
front of the allocated list, are then quick to find when deallocating,
and then that memory can be reused for similar objects later.